#93 [MOD] 코드 리뷰 반영
